backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.21.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.135.1",
    "google-adk>=1.26.0",
    "google-genai>=1.66.0",
    "motor>=3.7.1",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
    "redis>=7.4.0",
    "uvicorn>=0.41.0",
]
